vqa-svc/requirements.txt
flask-restplus==0.12.1
pyOpenSSL==19.0.0
git+ssh://git@github.com/BoseCorp/cortex-utils.git
flask-httpauth
requests
gunicorn
boto3
pytest
pytest-cov
psycopg2
